refactor(transform): share expression helpers (#122)
